# src/main/java/com/course/dto/LoginRequest.java
package com.course.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class LoginRequest {

    @JsonProperty
    private String username;

    @JsonProperty
    private String password;
}