# src/main/java/com/course/dto/MetadataRequest.java
package com.course.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@AllArgsConstructor
@NoArgsConstructor
@Setter
@Getter
public class MetadataRequest {

    @JsonProperty
    private String username;

    @JsonProperty
    private String metadata;

}